Edited value prop text and social icons.
Edited the value prop texts and the tagline. I found out if we can keep the value prop in two lines in the browser (in the largest browser size), it will look good better both in the browser and on mobile. Also, I put the social icons the center of the footer to match it with the top icons.